refactor: split api/bundles/bundles.py to multiple files
