+++ api/bundles/bundles.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 from typing import Literal
 
 from fastapi import Header, Query
@@ -15,96 +14,15 @@ from ayon_server.exceptions import (
     NotFoundException,
 )
 from ayon_server.lib.postgres import Postgres
-from ayon_server.types import NAME_REGEX, Field, OPModel, Platform
+from ayon_server.types import Field, OPModel, Platform
 
+from .actions import promote_bundle
+from .models import BundleModel, BundlePatchModel, ListBundleModel
 from .router import router
 
-dependency_packages_meta = {
-    "title": "Dependency packages",
-    "description": "mapping of platform:dependency_package_filename",
-    "example": {
-        "windows": "a_windows_package123.zip",
-        "linux": "a_linux_package123.zip",
-        "darwin": "a_mac_package123.zip",
-    },
-}
-
-
-class BaseBundleModel(OPModel):
-    pass
-
-
-class AddonDevelopmentItem(OPModel):
-    enabled: bool = Field(
-        True, example=False, description="Enable/disable addon development"
-    )
-    path: str = Field(
-        "", example="/path/to/addon", description="Path to addon directory"
-    )
-
-
-class BundleModel(BaseBundleModel):
-    """
-    Model for GET and POST requests
-    """
-
-    name: str = Field(
-        ...,
-        title="Name",
-        description="Name of the bundle",
-        example="my_superior_bundle",
-        regex=NAME_REGEX,
-    )
-
-    created_at: datetime = Field(
-        default_factory=datetime.now,
-        example=datetime.now(),
-    )
-    installer_version: str | None = Field(None, example="1.2.3")
-    addons: dict[str, str | None] = Field(
-        default_factory=dict,
-        title="Addons",
-        example={"ftrack": "1.2.3"},
-    )
-    dependency_packages: dict[Platform, str | None] = Field(
-        default_factory=dict, **dependency_packages_meta
-    )
-    is_production: bool = Field(False, example=False)
-    is_staging: bool = Field(False, example=False)
-    is_archived: bool = Field(False, example=False)
-    is_dev: bool = Field(False, example=False)
-    active_user: str | None = Field(None, example="admin")
-    addon_development: dict[str, AddonDevelopmentItem] = Field(
-        default_factory=dict,
-        example={"ftrack": {"enabled": True, "path": "~/devel/ftrack"}},
-    )
-
-
-class BundlePatchModel(BaseBundleModel):
-    addons: dict[str, str | None] = Field(
-        default_factory=dict,
-        title="Addons",
-        description="Changing addons is available only for dev bundles or server addons",
-        example={"ftrack": None, "kitsu": "1.2.3"},
-    )
-    installer_version: str | None = Field(None, example="1.2.3")
-    dependency_packages: dict[Platform, str | None] = Field(
-        default_factory=dict,
-        **dependency_packages_meta,
-    )
-    is_production: bool | None = Field(None, example=False)
-    is_staging: bool | None = Field(None, example=False)
-    is_archived: bool | None = Field(None, example=False)
-    is_dev: bool | None = Field(None, example=False)
-    active_user: str | None = Field(None, example="admin")
-    addon_development: dict[str, AddonDevelopmentItem] = Field(default_factory=dict)
-
-
-class ListBundleModel(OPModel):
-    bundles: list[BundleModel] = Field(default_factory=list)
-    production_bundle: str | None = Field(None, example="my_superior_bundle")
-    staging_bundle: str | None = Field(None, example="my_superior_bundle")
-    dev_bundles: list[str] = Field(default_factory=list)
+#
+# List all bundles
+#
 
 
 @router.get("/bundles", response_model_exclude_none=True)
@@ -118,26 +36,24 @@ async def list_bundles(
     dev_bundles: list[str] = []
 
     async for row in Postgres.iterate("SELECT * FROM bundles ORDER by created_at DESC"):
-        # postgres row is immutable, so let's make a copy
-        data = {**row["data"]}
+        # do not show archived bundles unless requested
+        if not archived and row["is_archived"]:
+            continue
 
-        # Clean-up in case there's a mess from previous versions
-        data.pop("is_production", None)
-        data.pop("is_staging", None)
-        data.pop("is_archived", None)
-        data.pop("active_user", None)
-        data.pop("is_dev", None)
+        data = row["data"]
 
-        # Construct the bundle model
         bundle = BundleModel(
-            **data,
             name=row["name"],
             created_at=row["created_at"],
+            addons=data.get("addons", {}),
+            installer_version=data.get("installer_version"),
+            dependency_packages=data.get("dependency_packages", {}),
             is_production=row["is_production"],
             is_staging=row["is_staging"],
             is_archived=row["is_archived"],
             is_dev=row["is_dev"],
             active_user=row["active_user"],
+            addon_development=data.get("addon_development", {}),
         )
 
         # helper top-level attributes (for convenience not crawling the list)
@@ -147,10 +63,6 @@ async def list_bundles(
             staging_bundle = row["name"]
         if row["is_dev"]:
             dev_bundles.append(row["name"])
-
-        # do not show archived bundles unless requested
-        if not archived and bundle.is_archived:
-            continue
 
         result.append(bundle)
 
@@ -162,7 +74,12 @@ async def list_bundles(
     )
 
 
-async def create_bundle(
+#
+# Create a new bundle
+#
+
+
+async def _create_new_bundle(
     bundle: BundleModel,
     user: UserEntity | None = None,
     sender: str | None = None,
@@ -252,13 +169,18 @@ async def create_new_bundle(
                 if addon_definition.latest:
                     bundle.addons[system_addon_name] = addon_definition.latest.version
 
-    await create_bundle(bundle, user, x_sender)
+    await _create_new_bundle(bundle, user, x_sender)
 
     return EmptyResponse(status_code=201)
 
 
+#
+# Update a bundle
+#
+
+
 @router.patch("/bundles/{bundle_name}", status_code=204)
-async def patch_bundle(
+async def update_bundle(
     bundle_name: str,
     bundle: BundlePatchModel,
     user: CurrentUser,
@@ -270,6 +192,7 @@ async def patch_bundle(
     ),
     x_sender: str | None = Header(default=None),
 ) -> EmptyResponse:
+
     if not user.is_admin:
         raise ForbiddenException("Only admins can patch bundles")
 
@@ -416,6 +339,11 @@ async def patch_bundle(
     return EmptyResponse(status_code=204)
 
 
+#
+# Delete bundle
+#
+
+
 async def delete_bundle(bundle_name: str):
     await Postgres.execute("DELETE FROM bundles WHERE name = $1", bundle_name)
 
@@ -431,71 +359,13 @@ async def delete_existing_bundle(
     return EmptyResponse(status_code=204)
 
 
+#
+# Bundle actions
+#
+
+
 class BundleActionModel(OPModel):
     action: Literal["promote"] = Field(..., example="promote")
-
-
-async def promote_bundle(bundle: BundleModel, user: UserEntity, conn):
-    """Promote a bundle to production.
-
-    That includes copying staging settings to production.
-    """
-
-    if not user.is_admin:
-        raise ForbiddenException("Only admins can promote bundles")
-
-    if not bundle.is_staging:
-        raise BadRequestException("Only staging bundles can be promoted")
-
-    if bundle.is_dev:
-        raise BadRequestException("Dev bundles cannot be promoted")
-
-    await conn.execute("UPDATE bundles SET is_production = FALSE")
-    await conn.execute(
-        """
-        UPDATE bundles
-        SET is_production = TRUE
-        WHERE name = $1
-        """,
-        bundle.name,
-    )
-
-    # Get project list
-    # statement = await conn.prepare("SELECT name FROM projects")
-    # project_names = [row["name"] async for row in statement.cursor()]
-
-    # Copy staging settings to production
-
-    for addon_name, addon_version in bundle.addons.items():
-        sres = await conn.fetch(
-            """
-                SELECT data FROM settings
-                WHERE addon_name = $1 AND addon_version = $2
-                AND variant = 'staging'
-                """,
-            addon_name,
-            addon_version,
-        )
-        if not sres:
-            data = {}
-        else:
-            data = sres[0]["data"]
-        await conn.execute(
-            """
-            INSERT INTO settings (addon_name, addon_version, variant, data)
-            VALUES ($1, $2, 'production', $3)
-            ON CONFLICT (addon_name, addon_version, variant)
-            DO UPDATE SET data = $3
-            """,
-            addon_name,
-            addon_version,
-            data,
-        )
-
-        # Do the same for every active project settings
-        # TODO: Do we want this?
-        #
-        # for project_name in project_names:
 
 
 @router.post("/bundles/{bundle_name}", status_code=201)
