+++ api/projects/list_projects.py
@@ -23,7 +23,9 @@ class ListProjectsItemModel(OPModel):
 class ListProjectsResponseModel(OPModel):
     detail: str = Field("OK", example="Showing LENGTH of COUNT projects")
     count: int = Field(
-        0, description="Total count of projects (regardless the pagination)", example=1,
+        0,
+        description="Total count of projects (regardless the pagination)",
+        example=1,
     )
     projects: list[ListProjectsItemModel] = Field(
         [],
@@ -44,7 +46,8 @@ class ListProjectsResponseModel(OPModel):
 async def list_projects(
     user: CurrentUser,
     page: int = Query(1, title="Page", ge=1),
-    length: int | None = Query(
+    length: int
+    | None = Query(
         None,
         title="Records per page",
         description="If not provided, the result will not be limited",
